updating to use shared helper to validate eval ready
